+++ CHANGELOG.md
@@ -2,6 +2,35 @@
 
 All notable changes to `laravel-translatable` will be documented in this file
 
+## 4.1.1 - 2019-02-27
+
+- fix service provider error
+
+## 4.1.0 - 2019-02-27
+
+- drop support for Laravel 5.7 and below
+- drop support for PHP 7.1 and below
+
+## 4.0.0 - 2019-02-27
+
+- `app.fallback_local` will now be used (see #148)
+
+## 3.1.3 - 2019-02-27
+
+- add support for Laravel 5.8
+
+## 3.1.2 - 2019-01-05
+
+- add `hasTranslation`
+
+## 3.1.1 - 2018-12-18
+
+- allow 0 to be used as a translation value
+
+## 3.1.0 - 2018-11-29
+
+- allow `getTranslations` to return other things than strings
+
 ## 3.0.1 - 2018-09-18
 
 - fix regarding empty locales